# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI Shield Gateway API",
    "description": "Zero-Leak PII Surrogate Shield, Drop-In OpenAI Privacy Proxy, Web Extractor & Cryptographic Grounding Verifier.",
    "version": "1.1.0",
    "contact": {
      "name": "Simultech AI Security",
      "url": "https://github.com/Simultech369"
    }
  },
  "servers": [
    {
      "url": "https://ai-shield-gateway.up.railway.app",
      "description": "Production Cloud Deployment"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health Probe",
        "description": "Returns API service health, uptime, and version.",
        "responses": {
          "200": {
            "description": "Service is healthy"
          }
        }
      }
    },
    "/v1/chat/completions": {
      "post": {
        "summary": "OpenAI Drop-In Privacy Proxy",
        "description": "Drop-in OpenAI proxy. Automatically sanitizes outbound messages by bijectively masking PII/secrets, forwards to upstream LLM, and unmasks response seamlessly before returning.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["messages"],
                "properties": {
                  "model": {
                    "type": "string",
                    "example": "gpt-4o"
                  },
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": { "type": "string", "example": "user" },
                        "content": { "type": "string", "example": "Send payment to 0x1111111111111111111111111111111111111111 and email bob@acme.org" }
                      }
                    }
                  },
                  "temperature": {
                    "type": "number",
                    "example": 0.7
                  },
                  "provider": {
                    "type": "string",
                    "example": "openai"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Standard OpenAI chat completion response with unmasked content and cryptographic PII audit receipt"
          },
          "402": {
            "description": "Payment Required via RapidAPI or Base USDC x402 protocol"
          }
        }
      }
    },
    "/v1/proxy": {
      "post": {
        "summary": "LLM Privacy Proxy Alias",
        "description": "Flexible alias for /v1/chat/completions supporting prompt strings or messages arrays across OpenAI, Groq, and Ollama.",
        "responses": {
          "200": {
            "description": "Unmasked completion with receipt"
          }
        }
      }
    },
    "/v1/mask": {
      "post": {
        "summary": "Mask PII & Secrets",
        "description": "Bijectively masks emails, API keys, private keys, ETH/crypto addresses, IP addresses, and paths with deterministic surrogates.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "example": "Contact me at alice@company.com with private key sk-abcdef12345678901234."
                  },
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sanitized text with cryptographic receipt and encrypted recovery token"
          },
          "402": {
            "description": "Payment Required via RapidAPI subscription or x402 Base USDC"
          }
        }
      }
    },
    "/v1/unmask": {
      "post": {
        "summary": "Unmask Completion",
        "description": "Re-hydrates original PII entities into LLM completion output using the recovery token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text", "recovery_token"],
                "properties": {
                  "text": {
                    "type": "string",
                    "example": "Sending email to {{SURROGATE_EMAIL_1}}."
                  },
                  "recovery_token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Re-hydrated plaintext"
          },
          "400": {
            "description": "Invalid recovery token"
          }
        }
      }
    },
    "/v1/extract": {
      "post": {
        "summary": "Web-to-Markdown AI Agent Extractor",
        "description": "Scrapes a public URL, strips ads, navs, and boilerplate, and outputs clean, token-efficient Markdown with structured metadata and SSRF protection.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url"],
                "properties": {
                  "url": {
                    "type": "string",
                    "example": "https://news.ycombinator.com"
                  },
                  "options": {
                    "type": "object",
                    "properties": {
                      "max_chars": { "type": "number", "example": 50000 },
                      "include_images": { "type": "boolean", "example": false }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Extracted Markdown and metadata with SHA-256 hash"
          },
          "403": {
            "description": "SSRF Blocked: Internal or dangerous host"
          }
        }
      }
    },
    "/v1/verify-citations": {
      "post": {
        "summary": "Verify Citations",
        "description": "Cryptographically verifies cited quotes against local corpus files or context packets, preventing RAG hallucinations.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "claim_id": {
                    "type": "string"
                  },
                  "citations": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "context_packets": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Grounding verification receipt with SHA-256 evidence"
          }
        }
      }
    },
    "/v1/scan-contract": {
      "post": {
        "summary": "Solidity Invariant & Security Linter",
        "description": "Performs static security analysis on Solidity smart contracts, detecting tx.origin vulnerabilities, reentrancy vectors, precision loss, unchecked transfers, and dangerous opcodes with cryptographic receipts.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["code"],
                "properties": {
                  "code": {
                    "type": "string",
                    "example": "pragma solidity ^0.8.20;\ncontract Vault { function withdraw() public { selfdestruct(payable(msg.sender)); } }"
                  },
                  "options": {
                    "type": "object",
                    "properties": {
                      "strict": { "type": "boolean", "example": true }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Security evaluation with list of flagged issues, security score (0-100), posture verdict, and signed SHA-256 receipt"
          },
          "400": {
            "description": "Missing code string"
          }
        }
      }
    }
  }
}